create turn end button component
